requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
httpx
openai
psycopg2-binary  # For PostgreSQL, or use 'sqlite' for SQLite
pytest
pytest-asyncio
